requirements.txt
psycopg2-binary == 2.8.6
django == 3.1.3
djangorestframework == 3.12.2
django-oauth-toolkit == 1.3.3
djangorestframework-simplejwt == 4.6.0
django-cors-headers == 3.5.0
django-filter == 2.4.0
django-taggit == 1.3.0
django-taggit-serializer == 0.1.7
django-extensions == 3.1.0
gunicorn==20.0.4
pydot == 1.4.1